Feature/diego

+++ bono.py
@@ -1,7 +1,16 @@
-# bono.py
+# ------ funciond calcular bono -----
 
-def calcular_bono(sueldo_base: int, porcentaje: float) -> float:
+def calcular_bono(sueldo_base: float, porcentaje: float) -> float:
   
     bono = sueldo_base * (porcentaje / 100)
     total = sueldo_base + bono
     return total
+
+
+# bono.py
+
+# def calcular_bono(sueldo_base: int, porcentaje: float) -> float:
+  
+#     bono = sueldo_base * (porcentaje / 100)
+#     total = sueldo_base + bono
+#     return total
